[rom_ext] encapsulate dice logics for moving to silicon_creator lib
This refactors the DICE-related logics in rom_ext.
Global variables are encapsulated in `dice_cert_t` and related code are encapsulated in `dice_chain_*` functions.

It also includes more boundary checks on the cert buffer in RAM to avoid overflow.

In order to have a better diff for code review, those changes are kept in rom_ext.c for now,
and we’ll move them to `silicon_creator/lib/cert/` in a follow-up PR.

This partially fixes #24368.

Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>