[crypto,cov] Add quit command to cryptotest firmware
This change introduces a `quit` command to the `cryptotest` firmware,
allowing the host to gracefully terminate device testing.

This ensures that coverage data is successfully reported back to the host upon
completion.

The `cryptotest` command enum and its handling function have been updated, and
the `aes_nist_kat` host test now sends the `quit` command and waits for coverage
data.

Change-Id: If984dab487905a5d4ff33c2b40ae86b0d8bb0249
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>