[cov] Refactor instrumented ROM loader
This change refactors the instrumented ROM loader to be a callable function
rather than a preprocessor conditional block.

This allows the feature to be enabled/disabled at bazel deps level.

Additionally, this change replaces the implicit Flash protection
configuration of the instrumented ROM region with an explicit one, fixing
tests with non-default Flash configuration in OTP settings.

Change-Id: I4b794c8aa9fdcc2dc294cbda804f01339c544545
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>