[sw] Add dual-redundancy boot_data writing algorithm
This commit introduces a new dual-redundancy scheme for writing boot
data to flash. Instead of the legacy circular writing scheme, which
is now renamed to `boot_data_write_old`, the new `boot_data_write`
function ensures that a valid and consistent entry is written to
the first index of both boot data pages.

Key changes:
- Renamed the legacy `boot_data_write` to `boot_data_write_old`.
- Implemented the new `boot_data_write` using a dual-redundancy
  algorithm.
- Added `boot_data_redundancy_check` to verify consistency between
  the two pages.
- Updated `rom_ext_start` to automatically repair boot data using
  the redundancy check.

Change-Id: Ic31d8c3c778397feb8eb2991ca9e670917047633
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>